# src/main/java/ru/skypro/homework/dto/CreateAdsDto.java
package ru.skypro.homework.dto;

import lombok.Data;

@Data
public class CreateAdsDto {

    private Integer pk;
    private String title;
    private String description;
    private String image;
    private Integer price;

}