# src/main/java/ru/skypro/homework/dto/CreateUserDto.java
package ru.skypro.homework.dto;

import lombok.Data;

@Data
public class CreateUserDto {

    private String email;
    private String password;
    private String firstName;
    private String lastName;
    private String phone;

}