# src/main/java/ru/skypro/homework/dto/UserDto.java
package ru.skypro.homework.dto;

import lombok.Data;

@Data
public class UserDto {

    private Integer id;
    private String email;
    private String firstName;
    private String lastName;
    private String phone;

}